% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegBuckner-package.R
\docType{package}
\name{ggsegBuckner-package}
\alias{ggsegBuckner}
\alias{ggsegBuckner-package}
\title{ggsegBuckner: Buckner Cerebellar Functional Parcellation for the 'ggseg' Ecosystem}
\description{
Buckner cerebellar parcellation into 7 or 17 functional networks for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegBuckner}
  \item Report bugs at \url{https://github.com/ggseg/ggsegBuckner/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
