% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buckner7}
\alias{buckner7}
\title{Buckner 7-Network Cerebellar Parcellation}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\usage{
data(buckner7)
}
\description{
Cerebellar parcellation into 7 functional networks from
Buckner et al. (2011). Contains 2D polygon geometry and 3D meshes.
}
\examples{
data(buckner7)
buckner7
}
\references{
Buckner RL et al. (2011). The organization of the human
cerebellum estimated by intrinsic functional connectivity.
\doi{10.1152/jn.00339.2011}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner17}}
}
\concept{ggseg_atlases}
\keyword{datasets}
